requirements.txt
requests
python-socketio[client]
nltk